requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
audioread==3.0.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langsmith==0.0.87
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
marshmallow==3.20.2
msgpack==1.0.7
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
numba==0.59.0
numpy==1.26.3
openai==1.11.1
packaging==23.2
pandas==2.2.0
platformdirs==4.2.0
pooch==1.8.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.26
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
websockets==12.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
